refactor code that shouldve been refactored a long time ago, fix crashes!
