+++ src/react/AppStatusProvider.tsx
@@ -1,8 +1,6 @@
 import { proxy, useSnapshot } from 'valtio'
-import { useEffect, useRef, useState } from 'react'
-import { activeModalStack, activeModalStacks, hideModal, insertActiveModalStack, miscUiState, showModal } from '../globalState'
-import { resetLocalStorageWorld } from '../browserfs'
-import { fsState } from '../loadSave'
+import { useEffect, useState } from 'react'
+import { activeModalStack, activeModalStacks, hideModal, insertActiveModalStack, miscUiState } from '../globalState'
 import { guessProblem } from '../errorLoadingScreenHelpers'
 import type { ConnectOptions } from '../connect'
 import { downloadPacketsReplay, packetsReplaceSessionState, replayLogger } from '../packetsReplay'
@@ -12,10 +10,9 @@ import DiveTransition from './DiveTransition'
 import { useDidUpdateEffect } from './utils'
 import { useIsModalActive } from './utilsApp'
 import Button from './Button'
-import { AuthenticatedAccount, updateAuthenticatedAccountData, updateLoadedServerData } from './ServersListProvider'
+import { updateAuthenticatedAccountData, updateLoadedServerData, AuthenticatedAccount } from './serversStorage'
 import { showOptionsModal } from './SelectOption'
 import LoadingChunks from './LoadingChunks'
-import MessageFormatted from './MessageFormatted'
 import MessageFormattedString from './MessageFormattedString'
 
 const initialState = {
